+++ hix/default.nix
@@ -1,26 +1,25 @@
+{ pkgs }:
 let
-  sources = import (../nix/sources.nix) {};
-  nixpkgs = import sources.nixpkgs-2009 {};
   args = '' --arg userDefaults "$HOME/.config/hix/hix.conf" --arg src ./.'';
   # Use HIX_ROOT to override the version of hix used when developing new hix features.
   # See docs/dev/hix.md for details.
   hixProject = "\${HIX_ROOT:-${./..}}/hix/project.nix";
-in nixpkgs.symlinkJoin {
+in pkgs.symlinkJoin {
   name = "hix";
   paths = [
-    (nixpkgs.writeScriptBin "hix-shell" ''
+    (pkgs.writeScriptBin "hix-shell" ''
       nix-shell ${hixProject} ${args} -A shell "$@"
     '')
-    (nixpkgs.writeScriptBin "hix-build" ''
+    (pkgs.writeScriptBin "hix-build" ''
       nix-build ${hixProject} ${args} "$@"
     '')
-    (nixpkgs.writeScriptBin "hix-instantiate" ''
+    (pkgs.writeScriptBin "hix-instantiate" ''
       nix-instantiate ${hixProject} ${args} "$@"
     '')
-    (nixpkgs.writeScriptBin "hix-env" ''
+    (pkgs.writeScriptBin "hix-env" ''
       nix-env -f ${hixProject} ${args} "$@"
     '')
-    (nixpkgs.writeScriptBin "hix" ''
+    (pkgs.writeScriptBin "hix" ''
       cmd=$1
       shift
       case $cmd in