+++ overlays/default.nix
@@ -1,4 +1,4 @@
-args:
+{ sources, ...}@args:
 
 let
   overlays = {
@@ -19,7 +19,7 @@ let
     nix-prefetch-git-minimal = import ./nix-prefetch-git-minimal.nix;
     gobject-introspection = import ./gobject-introspection.nix;
     hix = import ./hix.nix;
-    eval-packages = import ./eval-packages.nix;
+    eval-packages = import ./eval-packages.nix combined;
     ghcjs = import ./ghcjs.nix;
   };