+++ overlays/eval-packages.nix
@@ -1,4 +1,4 @@
-final: prev: {
+haskellNixOverlay: final: prev: {
   # This overlay makes `evalPackages` is like `buildPackages`, but on
   # `builtins.currentSystem` (when not building a nix flake).
   # We do not pass the `config` or `overlays` (not in `final.haskell-nix.overlays`).
@@ -6,11 +6,9 @@ final: prev: {
   # derivations that use `evalPackages` (these should be eval time only derviations
   # used to build nix inputs for IFD, the generated nix should match and so derivations
   # that depend on the IFD should match).
-  evalPackages =
+  evalPackages = (import final.path {
     # If we are building a flake there will be no currentSystem attribute
-    if builtins ? currentSystem
-      then (import final.path {
-          inherit (final.haskell-nix) overlays;
-        }).buildPackages
-      else final.buildPackages;
+    system = builtins.currentSystem or final.system;
+    overlays = [ haskellNixOverlay ];
+  }).buildPackages;
 }