+++ test/default.nix
@@ -1,4 +1,4 @@
-{ haskellNix ? import ../default.nix { inherit checkMaterialization; }
+{ haskellNix ? import ../default.nix { }
 , pkgs ? import nixpkgs nixpkgsArgs
 , nixpkgs ? haskellNix.sources.nixpkgs-unstable
 , nixpkgsArgs ? haskellNix.nixpkgsArgs
@@ -10,7 +10,7 @@
 with pkgs;
 
 let
-  inherit (import ../ci-lib.nix) dimension platformFilterGeneric filterAttrsOnlyRecursive;
+  inherit (import ../ci-lib.nix { inherit pkgs; }) dimension platformFilterGeneric filterAttrsOnlyRecursive;
   isDisabled = d: d.meta.disabled or false;
 
   # Set recurseForDerivations for both children and grand-children values in
@@ -165,25 +165,25 @@ let
     cabal-simple-debug = callTest ./cabal-simple-debug { inherit util compiler-nix-name; };
     cabal-simple-prof = callTest ./cabal-simple-prof { inherit util compiler-nix-name; };
     cabal-sublib = callTest ./cabal-sublib { inherit util compiler-nix-name; };
-    with-packages = callTest ./with-packages { inherit util; };
+    with-packages = callTest ./with-packages { inherit util compiler-nix-name; };
     builder-haddock = callTest ./builder-haddock {};
-    stack-simple = callTest ./stack-simple {};
-    stack-local-resolver = callTest ./stack-local-resolver {};
-    stack-local-resolver-subdir = callTest ./stack-local-resolver-subdir {};
-    stack-remote-resolver = callTest ./stack-remote-resolver {};
+    stack-simple = callTest ./stack-simple { inherit compiler-nix-name; };
+    stack-local-resolver = callTest ./stack-local-resolver { inherit compiler-nix-name; };
+    stack-local-resolver-subdir = callTest ./stack-local-resolver-subdir { inherit compiler-nix-name; };
+    stack-remote-resolver = callTest ./stack-remote-resolver { inherit compiler-nix-name; };
     shell-for-setup-deps = callTest ./shell-for-setup-deps { inherit compiler-nix-name; };
     setup-deps = import ./setup-deps { inherit pkgs compiler-nix-name; };
-    callStackToNix = callTest ./call-stack-to-nix {};
+    callStackToNix = callTest ./call-stack-to-nix { inherit compiler-nix-name; };
     callCabalProjectToNix = callTest ./call-cabal-project-to-nix { inherit compiler-nix-name; };
     cabal-source-repo = callTest ./cabal-source-repo { inherit compiler-nix-name; };
     cabal-source-repo-comments = callTest ./cabal-source-repo-comments { inherit compiler-nix-name; };
     buildable = callTest ./buildable { inherit compiler-nix-name; };
     project-flags-cabal = callTest ./project-flags/cabal.nix { inherit compiler-nix-name; };
-    project-flags-stack = callTest ./project-flags/stack.nix {};
+    project-flags-stack = callTest ./project-flags/stack.nix { inherit compiler-nix-name; };
     ghc-options-cabal = callTest ./ghc-options/cabal.nix { inherit compiler-nix-name; };
-    ghc-options-stack = callTest ./ghc-options/stack.nix {};
+    ghc-options-stack = callTest ./ghc-options/stack.nix { inherit compiler-nix-name; };
     exe-only = callTest ./exe-only { inherit util compiler-nix-name; };
-    stack-source-repo = callTest ./stack-source-repo {};
+    stack-source-repo = callTest ./stack-source-repo { inherit compiler-nix-name; };
     cabal-doctests = callTest ./cabal-doctests { inherit util compiler-nix-name; };
     extra-hackage = callTest ./extra-hackage { inherit compiler-nix-name; };
     hls-cabal = callTest ./haskell-language-server/cabal.nix { inherit compiler-nix-name; };
@@ -197,7 +197,7 @@ let
     coverage = callTest ./coverage { inherit compiler-nix-name; };
     coverage-golden = callTest ./coverage-golden { inherit compiler-nix-name;};
     coverage-no-libs = callTest ./coverage-no-libs { inherit compiler-nix-name; };
-    snapshots = callTest ./snapshots {};
+    snapshots = callTest ./snapshots { inherit compiler-nix-name; };
     sublib-docs = callTest ./sublib-docs { inherit util compiler-nix-name; };
     githash = haskell-nix.callPackage ./githash { inherit compiler-nix-name; testSrc = testSrcWithGitDir; };