+++ test/stack-simple/default.nix
@@ -1,4 +1,4 @@
-{ stdenv, lib, pkgs, mkStackPkgSet, haskellLib, testSrc }:
+{ stdenv, lib, pkgs, mkStackPkgSet, haskellLib, testSrc, compiler-nix-name }:
 
 with lib;
 
@@ -14,10 +14,12 @@ let
   packages = pkgSet.config.hsPkgs;
 
 in pkgs.recurseIntoAttrs {
+  meta.disabled = compiler-nix-name != "ghc865";
   stack-simple-exe = (haskellLib.check packages.stack-simple.components.exes.stack-simple-exe) // {
       # Attributes used for debugging with nix repl
       inherit pkgSet packages;
   };
   stack-simple-test = packages.stack-simple.checks.stack-simple-test;
   stack-simple-checks = packages.stack-simple.checks;
+  stack-simple-shell = packages.shellFor { tools = { cabal = "3.4.0.0"; }; };
 }