+++ test/unit.nix
@@ -6,23 +6,25 @@ let
        benchmarks = { };
        exes = { };
        foreignlibs = { };
-       library = "library";
+       library = { buildable = true; planned = true; };
        sublibs = { };
        tests = { };
     };
     package.identifier.name = "empty";
+    package.buildable = true;
   };
 
   componentsConfig = {
     components = {
-       benchmarks = { bbb = "bbb"; };
-       exes = { eee = "eee"; };
-       foreignlibs = { fff = "fff"; };
-       library = "library";
+       benchmarks = { bbb = { buildable = true; planned = true; }; };
+       exes = { eee = { buildable = true; planned = true; }; };
+       foreignlibs = { fff = { buildable = true; planned = true; }; };
+       library = { buildable = true; planned = true; };
        sublibs = { };
-       tests = { ttt = "ttt"; };
+       tests = { ttt = { buildable = true; planned = true; }; };
     };
     package.identifier.name = "nnn";
+    package.buildable = true;
   };
 
   testRepoData = {
@@ -65,7 +67,7 @@ lib.runTests {
     '');
     expected = __toJSON {
       otherText = "-- end of block\n";
-      sourceRepo = [(testRepoData // { subdir = "."; })];
+      sourceRepo = testRepoData // { subdirs = ["."]; };
     };
   };
 
@@ -80,7 +82,7 @@ lib.runTests {
     '');
     expected = __toJSON {
       otherText = "-- end of block\n";
-      sourceRepo = [(testRepoData // { subdir = "dir"; })];
+      sourceRepo = testRepoData // { subdirs = ["dir"]; };
     };
   };
 
@@ -95,7 +97,7 @@ lib.runTests {
     '');
     expected = __toJSON {
       otherText = "-- end of block\n";
-      sourceRepo = [(testRepoData // { subdir = "dir1"; }) (testRepoData // { subdir = "dir2"; })];
+      sourceRepo = testRepoData // { subdirs = ["dir1" "dir2"]; };
     };
   };
 
@@ -112,7 +114,7 @@ lib.runTests {
     '');
     expected = __toJSON {
       otherText = "-- end of block\n";
-      sourceRepo = [(testRepoData // { subdir = "dir1"; }) (testRepoData // { subdir = "dir2"; })];
+      sourceRepo = testRepoData // { subdirs = ["dir1" "dir2"]; };
     };
   };
 }